Move dedicated-worker code into content/src/html/workers/
Relocate the worker domain modules (worker, worker_global_scope,
dedicated_worker_global_scope, dedicated_worker_agent, worker_location,
worker_navigator) from content/src/html/ into a new workers/ subfolder,
and move the worker launcher subsystem (WorkerLauncher, WorkerRegistry,
ContentWorker) out of content/src/main.rs into
workers/worker_launcher.rs. html.rs now declares the workers module and
re-exports the Worker platform types unchanged; main.rs imports the
launcher and agent types from crate::html::workers. No behavior change.
