+++ content/src/html/workers/dedicated_worker_global_scope.rs
