+++ content/src/js/bindings/README.md
@@ -248,7 +248,7 @@ fn module_exports_binding<T: JsTypes>(
 | **Step comments on a JS binding function** | **Binding functions don't implement algorithms — no step comments** |
 | Prose doc comment on a non-spec helper (e.g. `simple_path`, `build_path_for_target`) | No doc comment at all — infrastructure functions have neither spec link nor prose |
 | Spec link on internal plumbing (reflector field, dispatch bridge, PostCreateReflector impl) | No doc comment — only code that directly implements a named spec algorithm or defines a spec concept gets an anchor |
-| Anchor of the algorithm that merely *calls* the helper (e.g. `#run-a-worker` on `start_script_fetch`, a fetch helper that run-a-worker step 12 invokes) | The anchor must claim what the item **is**. A helper that stands in for a named sub-algorithm the spec calls into is a partial implementation of that sub-algorithm and gets *its* anchor (`#fetch-a-classic-worker-script`), with the missing parts in body `// Note:`s — never the calling algorithm's anchor. Plumbing that implements no algorithm (IPC commands, enum variants, async continuations, state fields) carries no algorithm anchor; its doc comment names the exact step it serves. See `content/src/html/dedicated_worker_agent.rs::start_script_fetch` |
+| Anchor of the algorithm that merely *calls* the helper (e.g. `#run-a-worker` on `start_script_fetch`, a fetch helper that run-a-worker step 12 invokes) | The anchor must claim what the item **is**. A helper that stands in for a named sub-algorithm the spec calls into is a partial implementation of that sub-algorithm and gets *its* anchor (`#fetch-a-classic-worker-script`), with the missing parts in body `// Note:`s — never the calling algorithm's anchor. Plumbing that implements no algorithm (IPC commands, enum variants, async continuations, state fields) carries no algorithm anchor; its doc comment names the exact step it serves. See `content/src/html/workers/dedicated_worker_agent.rs::start_script_fetch` |
 | Putting a spec algorithm's anchor on a dictionary struct (e.g. `#event-flatten-more` on `AddEventListenerOptions`) | IDL dictionary types use `#dictdef-<name>` (e.g. `#dictdef-addeventlisteneroptions`); algorithms use `#event-`, `#concept-`, or `#dom-` prefixes |
 | Linking a dfn anchor from a function that implements an algorithm's steps (e.g. `#event-handler-content-attributes` on the handler-compilation function) | Anchor to the algorithm being implemented (e.g. `#getting-the-current-value-of-the-event-handler`). The dfn is the *concept*, not the steps. When the algorithm is unnamed in the spec (e.g. the attribute change steps synchronizing event handler content attributes), do NOT quote steps against a section link — no anchor shows those steps. Restructure the code so the named sub-algorithms carry the steps (e.g. `#deactivate-an-event-handler`, `#activate-an-event-handler`), and document the unnamed algorithm's function with body comments naming the algorithm instead of a link |
 | `JsObject` in a domain file (`content/src/dom/`, `content/src/html/`) | JsObject types belong in `content/src/js/` (the JS integration layer). Domain code operates on domain types like `EventTarget`, `Node`, `Event` |
